Lab 2 for Embded Systems Design 2019 Spring
A System instanciated by using a Qsys system. Assembly written using textmate and my brain. The main point of the system is to increment or decrement from 0-9 as well as rollover if the limit is met. Increment on switch0 is a 1 and the key is realsed, decrement when switch0 is a 0 and key released.